Move checkpointing module outside of helpers and rename to serialization
